Remove license kwarg
From the docs:

"The license field is a text indicating the license covering the package
where the license is not a selection from the “License” Trove
classifiers. See the Classifier field. Notice that there’s a licence
distribution option which is deprecated but still acts as an alias for
license."

The license is a selection from the "License" trove classifier, so this
kwarg is not needed.