Simplify index parsing
The ElementsBranch node has been removed, it can be represented
with the new WildcardIndex.  Previously foo[0] and foo[*] parsed
to two different structures.  Now both use a SubExpression node and
the wildcard is just represented as a WildcardIndex instead of an
Index node::

    SubExpression(AST(...), WildcardIndex())

Which makes the constructed ASTs more consistent.

I've also updated the parser grammar to not require separate rules
for root indices vs. indices in a subexpression.