Update spec with the JEP 6 grammar
Also removed the grammar from the README and replaced with a link.
Must people looking at the README don't care about the actual
grammar (other than to know that there is one).  For those that
are interested, they can follow the link to get to the grammar.
Also avoids us having to keep the grammar in two places.  Now the
official grammar lives in docs/specification.rst.