Add graphviz AST generator
This will take a parsed expression and generate a .dot file
that can be rendered via graphviz.  The sole purpose of this tool
is to enable devs to visualize the AST and help troubleshoot issues.
I've made this an internal interface for now, but only because the
AST is an implementation detail and subject to change.  People that
need this functionality are free to call the new method as needed.