Add support for using an ordered dictionary
This isn't part of the JMESPath spec (we use the JSON
types and a JSON object has no key order).  In order to
still accomodate this use case, I've added an Options class
that can alter how an expression is interpreted.  You can provide
a specific dict cls that should be used whenever the interpreter
creates a dict as the result of evaluating an expression.

This gives people that want this option a way to do it without
having to comply with the JMESPath spec.

Fixes #89