Always create a new copy of the function table
When defining a subclass of ``Functions``, the
``getattr()`` call would retrieve the base classes function
table, mutate the table, and then assign a reference of that
table to the child class.  We want to allow each subclass
to have its own separate copy of the function table so its
created from scratch each time.

I think there's room for optimization here.  We could
copy the parent func table if it exists and iterate through
the ``attrs``, which will only contain the functions defined
in the actual class.

Fixes #133.