Implement simpler fifo cache
This simplifies the caching logic while addressing the thread safety
issue when purging the cache.

Caching parsed expressions is essentially a historical artifact at this
point.  The parse times are on the order of microseconds now, so while
caching does improve performance, no one's going to notice microseconds
in Python.

As a compromise, the caching is left in place, with a simpler
implementation:

* A single element is purged if we exceed max cache size
* Bump the max cache size from 128 to 512
* Use basic FIFO eviction, relying on py37+ dict insertion order

Ran the failing test with 30 threads and `1_000_000` requests and
confirmed we don't see any errors and that the `self._MAX_SIZE` is
maintained.